Enable center shift mode in New Stickney layout.
The role of the Caps Lock key is changed as below:
* Caps Lock off: English mode
* Caps Lock on: Japanese mode

The default value for 'delay' configuration is changed to zero (no delay).

In New Stickney layout, the space bar is used as the center shift key.
Note by simply pressing and releasing the space bar, it is used as Henkan key.
These behavior is activated only in Japanese mode.

To type Katakana, just press the right shift key and enter the Katakana mode.
The Katakana mode is implicitly canceled when space bar is hit.

In the other layouts, to type Katakana, press Henkan key to enter the
Katakana mode, and press Muhenkan key to enter the Hiragana mode.

engine/Makefile.am
engine/bits.py
engine/engine.py
engine/new_stickney.py