Check if the surrounding text is updated after commit_text() strictly
On FC30, IBus.Capabilite.SURROUNDING_TEXT is always set with both Firefox and Chrome.
On Firefox, however, the surrounding text is not updated for committed text strings.
On Chrome, the surrounding text is never updated.
So we need to check we can use get_surrounding_text() or not more strictly.

engine/engine.py